Refactor tests
The tests were hardcoded, but now have helper methods for getting and
posting data to aid readability.
Note, however, that the 'get request' test remains trivial, as the value
it returns is hardcoded. This will be altered by posting data first, so
that there might be data that can be retrieved.